backend/requirements.txt
fastapi>=0.110,<0.112
uvicorn[standard]>=0.27,<0.28
python-multipart>=0.0.6
pydantic>=2.7,<3.0
pydantic-settings>=2.2,<3.0
SQLAlchemy>=2.0,<3.0
alembic>=1.13,<2.0
pypdf>=4.2,<5.0
tenacity>=8.2,<9.0
google-cloud-aiplatform>=1.45,<2.0
google-cloud-translate>=3.15
httpx>=0.27,<0.28
loguru>=0.7,<0.8
